requirements.txt
aioredis==1.2.0
amqp==2.4.1
asgiref==2.3.2
async-timeout==2.0.1
attrs==19.1.0
autobahn==19.2.1
Automat==0.7.0
billiard==3.5.0.5
celery==4.2.1
channels==2.1.7
channels-redis==2.3.2
constantly==15.1.0
daphne==2.2.5
Django==2.1.5
django-rest-auth==0.9.3
djangorestframework==3.9.1
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
kombu==4.3.0
msgpack==0.5.6
Pillow==5.4.1
psycopg2==2.7.7
PyHamcrest==1.9.0
pytz==2018.9
redis==3.2.0
six==1.12.0
Twisted==18.9.0
txaio==18.8.1
vine==1.2.0
zope.interface==4.6.0
